change data type for P_FactStrategyPlan procedure

+++ procedures/P_FactStrategyPlan.sql
@@ -12,7 +12,7 @@ CREATE PROC [dbo].[P_FactStrategyPlan]
 	@AdCost float,
 	@SpringAdCost float,
 	@BackSchool float,
-	@HolidayAdCost float,
+	@HolidayAdCost int,
 	@CostofGoodsSold float,
 	@SellCost float
 )
